+++ modules/iam-compartment/main.tf
@@ -6,7 +6,7 @@
 
 resource "oci_identity_compartment" "this" {
   count          = var.compartment_create ? 1 : 0
-  compartment_id = var.compartment_id
+  compartment_id = var.compartment_id != null ? var.compartment_id : var.tenancy_ocid
   name           = var.compartment_name
   description    = var.compartment_description
   enable_delete  = var.enable_delete
@@ -23,5 +23,6 @@ data "oci_identity_compartments" "this" {
 }
 
 locals {
-  compartment_ids = "${concat(flatten(data.oci_identity_compartments.this.*.compartments), list(map("id", "")))}"
+  compartment_ids        = concat(flatten(data.oci_identity_compartments.this.*.compartments), list(map("id", "")))
+  parent_compartment_ids = concat(flatten(data.oci_identity_compartments.this.*.compartments), list(map("compartment_id", "")))
 }